requirements.txt
dlib==19.10.0
halo==0.0.12
opencv_python==3.4.0.12
py_postgresql==1.2.1
typing==3.6.4
numpy==1.14.2
face_recognition==1.2.2
Pillow==6.1.0
PyYAML==5.1.1
